requirements.txt
alembic==1.13.3
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blinker==1.8.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.7
distro==1.9.0
fastapi==0.110.3
filelock==3.13.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.6
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
Mako==1.3.5
MarkupSafe==3.0.0
openai==1.51.1
outcome==1.3.0.post0
packaging==23.2
pbr==6.0.0
platformdirs==3.11.0
pluggy==1.5.0
psycopg2==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
PySocks==1.7.1
pytest==8.3.3
pytest-mock==3.14.0
python-dotenv==1.0.1
requests==2.31.0
selenium==4.17.2
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.35
starlette==0.37.2
stevedore==5.1.0
tqdm==4.66.5
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.1.0
uvicorn==0.29.0
virtualenv==20.24.6
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
webdriver-manager==4.0.1
Werkzeug==3.0.4
wsproto==1.2.0
